+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - MIT License with credits to packaged dependencies
+- Users can prefix all of Tailwind's styles with a configurable CSS selector.
 
 ## [0.1.0] - 2023-06-22